refactor(ModelEvent): refactor to utilize new traits for user, URL, and state management, enhancing modularity and code organization
